r846: Merge 2.1: Here is an awful lot of small and medium-sized changes, most of which are unrelated to each other and that cannot be reasonably attributed to some other larger topical changes to come.
This revision should match GIT::7a019f07a833ec9f598bc73c41c4f2253f5f845e.

+++ cinelerra/filegif.C
@@ -50,7 +50,7 @@ int64_t FileGIF::get_memory_usage()
 {
 // give buffer length plus padding
 	if(data)
-		return asset->width * asset->height * sizeof(VPixel);
+		return (int64_t)asset->width * asset->height * sizeof(VPixel);
 	else
 		return 256;
 }
